examples/web_scraper/requirements.txt
python-dotenv
openai
tanuki.py
pytest
beautifulsoup4